+++ src/client/actions/counter.js
@@ -1,0 +1,7 @@
+export const CHANGE = 'CHANGE';
+export function change(value) {
+  return {
+    type: CHANGE,
+    value,
+  };
+}