+++ src/client/actions/index.js
@@ -1,0 +1,5 @@
+import * as counter from './counter.js';
+
+export {
+  counter
+};