+++ src/client/index.jsx
@@ -1,9 +1,17 @@
 import 'antd/style/index.less';
-import App from './view/Counter';
+import App from './views/Counter';
 import ReactDOM from 'react-dom';
 import React from 'react';
+import { Provider } from 'react-redux';
+import configureStore from './store/configureStore';
+
+const store = configureStore();
 
 ReactDOM.render(
-  <App />,
+  <Provider store={store}>
+    <div>
+      <App />
+    </div>
+  </Provider>,
   document.getElementById('react-content')
 );